bot/requirements.txt
python-telegram-bot
openai
pinecone-client
python-dotenv
langchain